Revert "quelques erreurs en moins (7 errreurs symbole externes non résolus + 3 avertissements)"
This reverts commit 6f39a3e08b8959c467f4b8b384134ac7f748341f.

+++ Schotten_Totten/Schotten_Totten/VueJeu.cpp
@@ -18,7 +18,7 @@ void VueJeu::affichePlateau()
 		for (int ligne = maxJ1; ligne > -1; ligne--) {
 			for (unsigned int b = 0; b < 9; b++) {
 				vector<Carte*> cartesJ1 = (*bornes[b]).getCartesJ1();
-				if (cartesJ1.size() >= (static_cast<size_t>(ligne) + 1)) {
+				if (cartesJ1.size() >= (static_cast<long long>(ligne) + 1)) {
 					afficheCarte(cartesJ1[ligne]);
 				}
 				else { cout << "   "; }
@@ -37,7 +37,7 @@ void VueJeu::affichePlateau()
 		for (unsigned int ligne = 0; ligne < maxJ2; ligne++) {
 			for (unsigned int b = 0; b < 9; b++) {
 				vector<Carte*> cartesJ2 = (*bornes[b]).getCartesJ2();
-				if (cartesJ2.size() >= (static_cast<size_t>(ligne) + 1)) {
+				if (cartesJ2.size() >= (static_cast<long long>(ligne) + 1)) {
 					afficheCarte(cartesJ2[ligne]);
 				}
 				else { cout << "   "; }
